for #2567, move SelectStatement.getAggregationDistinctSelectItems to ShardingSelectOptimizedStatement
